Bump version and `cardano-graphql` dep to `1.0.0`, specify production release in changelog, improve README and deploy utility
