+++ CHANGELOG.md
@@ -2,4 +2,4 @@ Changelog
 =========
 
 ## 1.0.0
-First release
+First production-ready release.
