+++ README.md
@@ -5,6 +5,32 @@ Cardano Explorer App
 
 A React app with GraphQL client interfacing with [Cardano GraphQL](https://github.com/input-output-hk/cardano-graphql).
 
+### Environment Variables
+See [environment](source/environment.ts) for defaults.
+- `CARDANO_ERA`
+- `CARDANO_NETWORK`
+- `GRAPHQL_API_PROTOCOL`
+- `GRAPHQL_API_HOST`
+- `GRAPHQL_API_PORT`
+- `GRAPHQL_API_PATH`
+- `REAL_TIME_FACTOR`
+- `GA_TRACKING_ID`
+- `DEBUG`
+
+## Build
+```console
+yarn && yarn static:build
+```
+## Deploy
+The static bundle can be deployed using a standard web server. A simple [Node.js program](deploy/index.js) 
+is available for deploying the build to an AWS S3 bucket.
+
+```console
+AWS_ACCESS_KEY_ID=your_access_key_id \
+AWS_SECRET_ACCESS_KEY=your_secret_access_key node \
+./deploy/example_deployment.js
+```
+
 ## Development
 The environment is configured to access a remote managed deployment of the API, 
 however you can run a local stack using Docker and use a `.env` to work offline. 
@@ -12,7 +38,8 @@ See [.env.example](.env.example)
 
 ### `yarn dev`
 - Starts the development version of the app by default at http://localhost:4000
-- Generates graphql typings from the referenced schema in `cardano-graphql-ts` and documents within the codebase.
+- Generates graphql typings from the referenced schema in [`@cardano-graphql/client-ts`](https://github.com/input-output-hk/cardano-graphql/tree/master/packages/client-ts) 
+and documents within the codebase.
 - Any changes to graphql documents will trigger the TypeScript generator.
 
 ###  Storybook
@@ -28,15 +55,3 @@ The `develop` and PR branches are deployed continuously for the purpose of testi
 #### Testnet
 [![Netlify Status](https://api.netlify.com/api/v1/badges/16628b5d-b1f2-429b-a707-bbdec0564fe9/deploy-status)](https://cardano-explorer-testnet.netlify.app)
 
-## Deployment
-A simple [Node.js program](deploy/index.js) is available for deploying to an AWS S3 bucket.
-
-### ENVs
-\* optional
-#### `CARDANO_ERA`: `byron | shelley`
-#### `CARDANO_NETWORK`: `mainnet | testnet`
-#### `GRAPHQL_API_PROTOCOL` : `https | http`
-#### `GRAPHQL_API_HOST`
-#### `GRAPHQL_API_PORT`
-#### * `GA_TRACKING_ID`
-https://support.google.com/analytics/answer/7372977?hl=en
