+++ deploy/index.js
@@ -1,13 +1,14 @@
-const { spawn } = require('child_process');
-const upload = require('./upload');
+const fs = require('fs');
+const AWS = require('aws-sdk');
+const readdir = require('recursive-readdir');
+const mime = require('mime-types');
 
 module.exports = async function deploy() {
   const requiredEnvs = [
-    'CARDANO_ERA',
-    'CARDANO_NETWORK',
-    'GRAPHQL_API_PROTOCOL',
-    'GRAPHQL_API_HOST',
-    'GRAPHQL_API_PORT'
+    'AWS_ACCESS_KEY_ID',
+    'AWS_SECRET_ACCESS_KEY',
+    'AWS_DEFAULT_REGION',
+    'BUCKET'
   ];
 
   requiredEnvs.forEach(env => {
@@ -15,27 +16,27 @@ module.exports = async function deploy() {
     if (!val) {
       throw new Error(`Missing ${env} from the environment`);
     }
-
     console.log(`${env} = ${val}`);
   });
 
-  const builder = spawn(`yarn`, ['static:build'], {
-    stdio: ['inherit', 'inherit', 'inherit'],
-    env: {
-      PATH: process.env.PATH,
-      ...process.env
-    }
-  });
+  const targetBucket = process.env.BUCKET
 
-  await new Promise((resolve, reject) => {
-    builder.on('close', (code) => {
-      if (code !== 0) {
-        reject(new Error('Compilation failed'));
-      }
+  const uploadFolder = './build/static';
+  const s3 = new AWS.S3();
 
-      resolve();
-    })
-  });
+  const filesToUpload = await readdir(uploadFolder);
 
-  await upload();
+  console.log(`Uploading to S3 bucket: ${targetBucket}`);
+
+  return filesToUpload.map(async (file) => {
+    const key = file.replace('build/static/', '');
+    const fileBuffer = fs.readFileSync(file);
+
+    return s3.upload({
+      Key: key,
+      Bucket: targetBucket,
+      Body: fileBuffer,
+      ContentType: mime.lookup(key) || 'application/octet-stream'
+    }).promise();
+  })
 };
